datagen/gen2/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==0.17.0
    # via -r requirements.in
azure-core==1.11.0
    # via azure-storage-blob
azure-storage-blob==12.7.1
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   msrest
    #   requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.6
    # via azure-storage-blob
docopt==0.6.2
    # via -r requirements.in
faker==13.11.1
    # via -r requirements.in
flake8==3.8.4
    # via -r requirements.in
humanize==3.2.0
    # via -r requirements.in
idna==2.10
    # via requests
isodate==0.6.0
    # via msrest
jinja2==2.11.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
msrest==0.6.21
    # via azure-storage-blob
oauthlib==3.1.0
    # via requests-oauthlib
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pymongo==3.11.3
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   arrow
    #   faker
redis==3.5.3
    # via -r requirements.in
requests==2.25.1
    # via
    #   -r requirements.in
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
six==1.15.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
urllib3==1.26.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
